Make wrap keyword available in cubefromshape
Wrap is a keyword which is projection specific and which gives the
projection value at which the projection wraps into itself.
The default values fail for non Lat Lon projections.
Therefore, this makes the keyword available in the cubefromshape
interface.
This is an intermediate solution, because it would be better to have the
projection as a metadata information of a cube, so that we could use
this to set the wrapping values.